refactor(api): split the stats aggregation service per entity
The team and player queries share no SQL, no row shape and no helpers, so
they are now TeamStatsAggregationService and PlayerStatsAggregationService
and each file only pulls in the models its own query touches.

Also fixes an inconsistency the parallel test suites exposed: a cached
ordering can name an entity deleted since it was built, so the page came
back short while total still counted it. A short page is now treated as a
stale cache and the totals are recomputed once, keeping items and total
from the same snapshot. expectedPageSize distinguishes that from an
ordinary last page.
